working version

+++ app/src/main/java/ru/systempla/weatherapp/mvp/model/entity/ForecastEntityRestModel.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 class ForecastEntityRestModel {
     @SerializedName("dt")
     @Expose
-    var dt: Int? = null
+    var dt: Long? = null
 
     @SerializedName("main")
     @Expose
